# Cache Jekyll.sanitized_path