--- a/lib/jekyll.rb
+++ b/lib/jekyll.rb
@@ -175,6 +175,12 @@ module Jekyll
       return base_directory if base_directory.eql?(questionable_path)
       return base_directory if questionable_path.nil?
 
+      @sanitized_path_cache ||= {}
+      @sanitized_path_cache[base_directory] ||= {}
+      if @sanitized_path_cache[base_directory][questionable_path]
+        return @sanitized_path_cache[base_directory][questionable_path]
+      end
+
       clean_path = questionable_path.dup
       clean_path.insert(0, "/") if clean_path.start_with?("~")
       clean_path = File.expand_path(clean_path, "/")
@@ -185,12 +191,13 @@ module Jekyll
       # `File.expand_path` above.
       clean_path.squeeze!("/")
 
-      if clean_path.start_with?(base_directory.sub(%r!\z!, "/"))
-        clean_path
-      else
-        clean_path.sub!(%r!\A\w:/!, "/")
-        File.join(base_directory, clean_path)
-      end
+      @sanitized_path_cache[base_directory][questionable_path] =
+        if clean_path.start_with?(base_directory.sub(%r!\z!, "/"))
+          clean_path
+        else
+          clean_path.sub!(%r!\A\w:/!, "/")
+          File.join(base_directory, clean_path)
+        end
     end
 
     # Conditional optimizations