--- a/lib/jekyll.rb
+++ b/lib/jekyll.rb
@@ -172,6 +172,9 @@ module Jekyll
     #
     # Returns the sanitized path.
     def sanitized_path(base_directory, questionable_path)
+      return base_directory if base_directory.eql?(questionable_path)
+      return base_directory if questionable_path.nil?
+
       Jekyll::PathManager.sanitized_path(base_directory, questionable_path).dup
     end