--- a/lib/jekyll/path_manager.rb
+++ b/lib/jekyll/path_manager.rb
@@ -33,36 +33,30 @@ module Jekyll
     #
     # Returns a frozen string.
     def self.sanitized_path(base_directory, questionable_path)
-      cached_path = sanitized_path_cache(base_directory, questionable_path)
-      return cached_path if cached_path
+      @sanitized_path ||= {}
+      @sanitized_path[base_directory] ||= {}
+      @sanitized_path[base_directory][questionable_path] ||= begin
+        return base_directory.freeze if questionable_path.nil?
 
-      clean_path = questionable_path.dup
-      clean_path.insert(0, "/") if clean_path.start_with?("~")
-      clean_path = File.expand_path(clean_path, "/")
+        clean_path = if questionable_path.start_with?("~")
+                       questionable_path.dup.insert(0, "/")
+                     else
+                       questionable_path
+                     end
+        clean_path = File.expand_path(clean_path, "/")
+        return clean_path.freeze if clean_path.eql?(base_directory)
 
-      return clean_path.freeze if clean_path.eql?(base_directory)
+        # remove any remaining extra leading slashes not stripped away by calling
+        # `File.expand_path` above.
+        clean_path.squeeze!("/")
 
-      # remove any remaining extra leading slashes not stripped away by calling
-      # `File.expand_path` above.
-      clean_path.squeeze!("/")
-
-      cached_path =
         if clean_path.start_with?(base_directory.sub(%r!\z!, "/"))
           clean_path.freeze
         else
           clean_path.sub!(%r!\A\w:/!, "/")
-          Jekyll::PathManager.join(base_directory, clean_path)
+          join(base_directory, clean_path)
         end
-
-      sanitized_path_cache(base_directory, questionable_path, cached_path)
+      end
     end
-
-    def self.sanitized_path_cache(base_directory, questionable_path, cache_path = nil)
-      @sanitized_path ||= {}
-      @sanitized_path[base_directory] ||= {}
-      @sanitized_path[base_directory][questionable_path] ||= cache_path
-    end
-    private_class_method :sanitized_path_cache
-
   end
 end