--- a/test/test_path_manager.rb
+++ b/test/test_path_manager.rb
@@ -3,21 +3,31 @@
 require "helper"
 
 class TestPathManager < JekyllUnitTest
-  context "results are always frozen" do
+  context "PathManager" do
     setup do
       @source = Dir.pwd
     end
 
-    should "freeze the base directory" do
-      assert Jekyll::PathManager.sanitized_path(@source, @source).frozen?
+    should "return frozen copy of base if questionable path is nil" do
+      assert_equal @source, Jekyll::PathManager.sanitized_path(@source, nil)
+      assert Jekyll::PathManager.sanitized_path(@source, nil).frozen?
     end
 
-    should "freeze a path inside the base directory" do
-      assert Jekyll::PathManager.sanitized_path(@source, File.join(@source, "_config.yml")).frozen?
+    should "return a frozen copy of base if questionable path expands into the base" do
+      assert_equal @source, Jekyll::PathManager.sanitized_path(@source, File.join(@source, "/"))
+      assert Jekyll::PathManager.sanitized_path(@source, File.join(@source, "/")).frozen?
     end
 
-    should "freeze a relative path" do
-      assert Jekyll::PathManager.sanitized_path(@source, "_config.yml").frozen?
+    should "return a frozen string result" do
+      assert_equal(
+        "#{@source}/_config.yml",
+        Jekyll::PathManager.sanitized_path(@source, "E:\\_config.yml")
+      )
+      assert_equal(
+        "#{@source}/_config.yml",
+        Jekyll::PathManager.sanitized_path(@source, "//_config.yml")
+      )
+      assert Jekyll::PathManager.sanitized_path(@source, "//_config.yml").frozen?
     end
   end
 end