@ashmaroli 2020-10-05
Thank you for this submission, @fauno. I love optimization pull requests.
However, I have some reservations here though.

`Jekyll` is the global `module`. I'm not very confident about instance variables in a global module.
That aside, earlier the result could be freely mutated. Now since it is cached, mutating the resulting string in one place will affect result elewhere..