@fauno 2020-10-05
Thanks for the review! Would it be ok to move the cache to a private method in the module?  Since it's caching results I don't find it so problematic to share the cache/instance var, since the result is idempotent anyway, no?

As per mutability, I didn't take that into account, it could return a duplicated string, the benchmark looks better:

| benchmark | with patch and dup result | with patch | without patch |
| --------- | ------------------------- | ---------- | ------------- |
| with no questionable path | 72.103k i/100ms | 69.426k i/100ms | 6.977k i/100ms |
| with a single-part questionable path | 73.170k i/100ms | 63.989k i/100ms | 6.620k i/100ms |
| with a multi-part questionable path | 72.014k i/100ms | 67.406k i/100ms | 6.350k i/100ms |
| with a single-part traversal path | 68.148k i/100ms | 68.679k i/100ms | 6.544k i/100ms |
| with a multi-part traversal path | 67.653k i/100ms | 63.054k i/100ms | 6.314k i/100ms |
| with the exact same paths | 384.003k i/100ms | 411.366k i/100ms | 402.619k i/100ms |
| with a single-part absolute path including the base_directory | 30.057k i/100ms | 28.353k i/100ms | 6.656k i/100ms |
| with a multi-part absolute path including the base_directory | 28.053k i/100ms | 29.346k i/100ms | 6.443k i/100ms |
| with a single-part traversal path including the base_directory | 9.830k i/100ms | 9.966k i/100ms | 12.621k i/100ms |
| with a multi-part traversal path including the base_directory | 43.127k i/100ms | 40.974k i/100ms | 5.848k i/100ms |
| with no questionable path | 740.184k (± 1.4%) i/s -      3.749M in   5.066490s | 674.568k (± 2.8%) i/s -      3.402M in   5.047108s | 70.152k (± 1.9%) i/s -    355.827k in   5.074255s |
| with a single-part questionable path | 731.445k (± 0.7%) i/s -      3.658M in   5.002016s | 656.353k (± 4.1%) i/s -      3.327M in   5.079440s | 66.758k (± 2.7%) i/s -    337.620k in   5.061671s |
| with a multi-part questionable path | 686.939k (±12.2%) i/s -      3.385M in   5.055924s | 642.866k (± 3.0%) i/s -      3.235M in   5.037611s | 63.044k (± 4.2%) i/s -    317.500k in   5.046474s |
| with a single-part traversal path | 722.345k (± 1.4%) i/s -      3.612M in   5.001173s | 655.010k (± 2.8%) i/s -      3.297M in   5.037056s | 66.523k (± 2.6%) i/s -    333.744k in   5.020730s |
| with a multi-part traversal path | 659.161k (± 5.8%) i/s -      3.315M in   5.049436s | 618.577k (± 2.9%) i/s -      3.153M in   5.101243s | 59.564k (±17.5%) i/s -    290.444k in   5.133958s |
| with the exact same paths | 3.821M (± 2.8%) i/s -     19.200M in   5.028773s | 3.771M (± 9.5%) i/s -     18.923M in   5.070652s | 3.384M (±19.7%) i/s -     16.507M in   5.128966s |
| with a single-part absolute path including the base_directory | 300.236k (± 1.4%) i/s -      1.503M in   5.006624s | 285.799k (± 3.5%) i/s -      1.446M in   5.066518s | 52.928k (±20.3%) i/s -    252.928k in   5.009928s |
| with a multi-part absolute path including the base_directory | 281.253k (± 1.5%) i/s -      1.431M in   5.088109s | 282.806k (± 5.1%) i/s -      1.438M in   5.101322s | 57.189k (±21.4%) i/s -    277.049k in   5.216751s |
| with a single-part traversal path including the base_directory | 101.875k (± 1.8%) i/s -    511.160k in   5.019353s | 97.885k (± 1.6%) i/s -    498.300k in   5.092095s | 82.299k (±22.6%) i/s -    391.251k in   5.068463s |
| with a multi-part traversal path including the base_directory | 429.309k (± 1.7%) i/s -      2.156M in   5.024406s | 400.997k (± 2.5%) i/s -      2.008M in   5.010275s | 37.547k (±12.6%) i/s -    187.136k in   5.087265s |