@ashmaroli 2020-10-05
Thanks for the numbers! Very inspiring!

> since the result is idempotent anyway, no?

In a way, yes this is true.
But since this is Ruby, things can get complicated.

For example consider the following hypothetical plugin:
```ruby
class Foo
  extend Jekyll

  @sanitization_path_cache = {}

  def self.clarify(input)
    @sanitization_path_cache[input] ||= begin
      lorem = call_business_logic(input)
      sanitized_path(base, lorem)
    end
  end
end
```
The code above would was as the author expected until Jekyll-4.1.
With Jekyll 4.2, *it may not be so..*
Additionally, `@sanitization_path_cache` in the plugin above need not be just a Hash but a specialized object in itself as well.

The reason I'm saying this is not to discourage you but to err or the safer side. In the past, I myself had used instance variables as cache in a module (see `Jekyll::Filters#where`). Luckily, only one user had problems because of it (to my knowledge).



> I didn't take that into account, it could return a duplicated string

Ah! That's a nice solution.
If you're returning a duplicated string, you could also replace `File.join(*)` with `Jekyll::PathManager.join(*)` for additional optimization.